refactor(proc): split into multiple files
